+++ packages/components/src/components/textarea/textarea-styles.ts
@@ -27,6 +27,11 @@ export const cssVarInternalTextareaScaling = '--p-internal-textarea-scaling';
 // CSS Variable defined in fontHyphenationStyle
 /**
  * @css-variable {"name": "--p-hyphens", "description": "Sets the CSS `hyphens` property for text elements, controlling whether words can break and hyphenate automatically.", "defaultValue": "auto"}
+ * @css-variable {"name":"--p-textarea-field-sizing","description":"Controls CSS `field-sizing` for textarea.","defaultValue":"unset"}
+ * @css-variable {"name":"--p-textarea-min-width","description":"Minimum width of the textarea.","defaultValue":"52px"}
+ * @css-variable {"name":"--p-textarea-max-width","description":"Maximum width of the textarea.","defaultValue":"unset"}
+ * @css-variable {"name":"--p-textarea-min-height","description":"Minimum height of the textarea.","defaultValue":"unset"}
+ * @css-variable {"name":"--p-textarea-max-height","description":"Maximum height of the textarea.","defaultValue":"unset"}
  */
 
 export const getComponentCss = (
@@ -63,6 +68,11 @@ export const getComponentCss = (
         gridArea: '1/1',
         display: 'block',
         resize,
+        fieldSizing: 'var(--p-textarea-field-sizing, unset)',
+        minWidth: `var(--p-textarea-min-width, 2ch)`,
+        maxWidth: 'var(--p-textarea-max-width, unset)',
+        minHeight: 'var(--p-textarea-min-height, unset)',
+        maxHeight: 'var(--p-textarea-max-height, unset)',
         border: `${borderWidthThin} solid ${formStateBorderColor}`,
         borderRadius: `var(${legacyRadiusSmall}, ${isCompact ? radiusLg : radiusXl})`,
         background: formStateBackgroundColor,