+++ packages/storefront/src/utils/majorVersion.ts
@@ -1,0 +1,6 @@
+import pkgJson from '../../package.json' with { type: 'json' };
+
+export const majorVersion = parseInt(
+  pkgJson.dependencies['@porsche-design-system/components-react'].replace(/^\^|^~/, ''),
+  10
+);